Revert "Yoma-442 - refactor(DigitalCV): my cv display real data in about widget"
