+++ src/components/CvCard/CvCard.tsx
@@ -8,7 +8,6 @@ import Text, { TextAlign } from '../Typography'
 import styles from './CvCard.styles'
 
 type Props = {
-  text: string
   cardTitle: string
   defaultText: string
   count?: number
@@ -18,7 +17,6 @@ type Props = {
 }
 
 const CvCard = ({
-  text,
   cardTitle,
   defaultText,
   count = 0,
@@ -36,7 +34,7 @@ const CvCard = ({
       </View>
       <View style={styles.dividerLine} />
       <View style={styles.bodyView}>
-        <Text.Body align={TextAlign.center}>{text || defaultText}</Text.Body>
+        <Text.Body align={TextAlign.center}>{defaultText}</Text.Body>
       </View>
     </View>
   )
