--- src/modules/DigitalCv/DigitalCv.selector.ts
@@ -1,7 +1,0 @@
-import { createSelector } from '@reduxjs/toolkit'
-
-import * as userSelectors from '../User/User.selector'
-
-export default createSelector(userSelectors.selectBiography, biography => ({
-  biography,
-}))
